Added a shop to the game
