+++ GameWIndowTest1/out_of_combat.xaml.cs
@@ -30,12 +30,6 @@ namespace GameWIndowTest1
         List<TextBlock> _character_blocks;
         List<ComboBox> ComboBoxes;
 
-        int heal_ammount = 10;
-        int heal_cost = 500;
-        int revive_cost = 1000;
-        int restore_uses_cost = 200;
-
-
         bool init_setup = true;
 
         int selected_index = 0;
@@ -56,36 +50,17 @@ namespace GameWIndowTest1
             //_characters[0].takedamage(15);
             //_characters[1].takedamage(8);
             set_character_details();
-
-            Mid_Block.Text = $"{state.money} money";
-            Mid_Block.Text += $"\nHeals cost {heal_cost} for {heal_ammount} healing";
-            Mid_Block.Text += $"\nRevives cost {revive_cost}";
-            Mid_Block.Text += $"\nRestore cost {restore_uses_cost}";
-            set_ability_selector_vals();
         }
 
         public void set_character_details()
         {
+            
+            if (_characters.Count == 0)
+            {
+                return;
+            }
+
             character _current = _characters[selected_index];
-            // if the selected character cannot be healed anymore
-            if (_current.IsDead || _current.health >= _current.max_health || state.money <= heal_cost)
-            {
-                Healing_Button.IsEnabled = false;
-            }
-            else
-            {
-                Healing_Button.IsEnabled = true;
-            }
-
-            if (_current.IsDead && state.money >= revive_cost)
-            {
-                Revive_Button.IsEnabled = true;
-            }
-            else
-            {
-                Revive_Button.IsEnabled = false;
-            }
-
 
 
             for (int index = 0; index<_characters.Count; index++)
@@ -104,7 +79,6 @@ namespace GameWIndowTest1
                 }
             }
 
-            List<ComboBox> ComboBoxes = new List<ComboBox> { Ability_Box1, Ability_Box2, Ability_Box3, Ability_Box4 };
             foreach (ComboBox box in ComboBoxes)
             {
                 // clear all the items in the current box
@@ -123,7 +97,6 @@ namespace GameWIndowTest1
 
         private void Next_fight(object sender, RoutedEventArgs e)
         {
-            convert_boxes_to_abilities();
             // increment the current wave number of the game
             state.current_wave_number++;
             MainWindow game_window = new MainWindow(state);
@@ -131,49 +104,21 @@ namespace GameWIndowTest1
             this.Close();
         }
 
+        private void Shop_Button_Click(object sender, RoutedEventArgs e)
+        {
+            shop show_Window= new shop(state, selected_index);
+            show_Window.Show();
+            this.Close();
+        }
+
         private void Radio_Changed(object sender, RoutedEventArgs e)
         {
             RadioButton rb = sender as RadioButton;
-
-            // update the last characters abilities
-            convert_boxes_to_abilities();
             // then change who is the selected character
-            selected_index = Int32.Parse(rb.Name.Substring(rb.Name.Length-1,1))-1;
+            selected_index = Int32.Parse(rb.Name.Substring(rb.Name.Length - 1, 1)) - 1;
 
-            if (_characters.Count == 0)
-            {
-                Healing_Button.IsEnabled = false;
-                return;
-            }
             set_character_details();
 
-        }
-
-        private void Heal_Button(object sender, RoutedEventArgs e)
-        {
-            Mid_Block.Text = $"\n{state.money-heal_cost} money";
-            Mid_Block.Text += $"\nHeals cost {heal_cost} for {heal_ammount} healing";
-            Mid_Block.Text += $"\nRevives cost {revive_cost}";
-            Mid_Block.Text += $"\nRestores cost {restore_uses_cost}";
-
-            character current = _characters[selected_index];
-            Mid_Block.Text += $"\n\n{current.display_name} had {current.health}";
-
-            Mid_Block.Text += $"\nHealed for {heal_ammount}";
-
-            // heal the character for that ammount
-            current.heal(heal_ammount);
-
-            Mid_Block.Text += $"\nand now has {current.health} health";
-
-            state.money -= heal_cost;
-
-            if (state.money <= heal_cost);
-            {
-                // disable the healing button as it has run out of uses
-                Healing_Button.IsEnabled = false;
-            }
-            set_character_details();
         }
 
         private void Save_Button_Click(object sender, RoutedEventArgs e)
@@ -192,34 +137,6 @@ namespace GameWIndowTest1
             File.WriteAllText(fileName, jsonString);
             MessageBox.Show("Saved");
 
-        }
-
-        private void Revive_Button_Click(object sender, RoutedEventArgs e)
-        {
-            character current = _characters[selected_index];
-            Mid_Block.Text = $"{state.money-revive_cost} money";
-            Mid_Block.Text += $"\nHeals cost {heal_cost} for {heal_ammount} healing";
-            Mid_Block.Text += $"\nRevives cost {revive_cost}";
-            Mid_Block.Text += $"\nRestores cost {restore_uses_cost}";
-
-            Mid_Block.Text += $"\n\n{current.display_name} had {current.health}";
-
-            // if the current it not dead
-            if (!current.IsDead)
-            {
-                return;
-            }
-
-            current.revive();
-
-            state.money -= revive_cost;
-
-            if (state.money <= revive_cost)
-            {
-                // disable the healing button as it has run out of uses
-                Revive_Button.IsEnabled = false;
-            }
-            set_character_details();
         }
 
         private void Ability_box_changed(object sender, SelectionChangedEventArgs e)
@@ -250,7 +167,6 @@ namespace GameWIndowTest1
                 MessageBox.Show($"{new_ability.name} is already equiped so cannot be equiped again");
                 set_last_item_box(box, current_character.abilities[ability_index]);
                 // and return so that the ability is not change
-                set_ability_selector_vals();
                 return;
             }
 
@@ -259,7 +175,6 @@ namespace GameWIndowTest1
             MessageBox.Show($"Ability {ability_index} changed from {current_character.abilities[ability_index].name} to {new_ability.name}");
             // and change it
             current_character.updateAbility(new_ability, ability_index);
-            set_ability_selector_vals();
        }
 
         public void set_last_item_box(ComboBox box, ability previous_ability)
@@ -321,61 +236,6 @@ namespace GameWIndowTest1
                 ComboBox box = ComboBoxes[i];
                 box.SelectedIndex = box.Items.IndexOf(current_character.abilities[i].name);
             }
-        }
-
-        public void set_ability_selector_vals()
-        {
-            character current_character = state.characters[selected_index];
-            ComboBox abilityBox = Ability_Selector;
-            abilityBox.Items.Clear();
-            for (int i = 0; i < current_character.abilities.Length; i++)
-            {
-                string current_ability = current_character.abilities[i].name;
-                //MessageBox.Show(current_ability.ToString());
-                string val = $"{current_ability}";
-                abilityBox.Items.Add(val);
-            }
-            abilityBox.SelectedIndex = 0;
-        }
-
-        private void Restore_Uses_Click(object sender, RoutedEventArgs e)
-        {
-            ComboBox abilityBox = Ability_Selector;
-            character current = _characters[selected_index];
-            Mid_Block.Text = $"{state.money - restore_uses_cost} money";
-            Mid_Block.Text += $"\nHeals cost {heal_cost} for {heal_ammount} healing";
-            Mid_Block.Text += $"\nRevives cost {revive_cost}";
-            Mid_Block.Text += $"\nRestores cost {restore_uses_cost}";
-
-            string ability_name = abilityBox.SelectedItem.ToString();
-            ability selectedAbility = character.no_ability_selected;
-
-            foreach (ability _a in current.abilities)
-            {
-                // if this is the ability that was selected by the player
-                if (_a.name == ability_name) { selectedAbility = _a; break; }
-            }
-
-            if (selectedAbility == character.no_ability_selected)
-            {
-                // if no ability was selected 
-                // return and cost nothing
-                return;
-            }
-
-            MessageBox.Show($"{selectedAbility.name}: {selectedAbility.uses_remaining}/{selectedAbility.max_number_of_uses}");
-
-            // set the number of uses remaing on this ability to the max number of uses it can have
-            selectedAbility.uses_remaining = selectedAbility.max_number_of_uses;
-
-            state.money -= restore_uses_cost;
-
-            if (state.money <= restore_uses_cost)
-            {
-                // disable the healing button as it has run out of uses
-                Restore_Uses.IsEnabled = false;
-            }
-            set_character_details();
         }
     }
 }
